requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
requests>=2.28.1
pillow>=9.0
slowapi>=0.1.6
pytest>=7.1.3
